Revert "맵 확장 밎 코인 애니메이션 작업"
This reverts commit cc7b3ecb59dac3f48c4184b9639cfcc595fd35a5.
