+++ Animon/Assets/CoinScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad9ba2baf660bf34ca1175acff1211de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
