--- Animon/Assets/Prefabs/Box.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a8c0849a8454be7428dad01782e0a77b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
