+++ Animon/Assets/Prefabs/Coin.prefab
@@ -29,7 +29,7 @@ Transform:
   m_GameObject: {fileID: 4286075480770937419}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 2, y: 1, z: 0}
-  m_LocalScale: {x: 0.5, y: 1, z: 1}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
@@ -110,4 +110,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ad9ba2baf660bf34ca1175acff1211de, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  rotationSpeed: 60
