--- Animon/Assets/Scripts/BoxSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b7aeea794b73c14ca394c2cb2dc1568
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
