--- Animon/Assets/Scripts/CameraController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 318a900d7b77bcb448b16164a4e8a6c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
