--- Animon/Assets/Scripts/CoinScript.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CoinScript : MonoBehaviour
-{
-    public float rotationSpeed = 60f;
-
-    void Update()
-    {
-        transform.Rotate(0f, rotationSpeed * Time.deltaTime, 0f);
-    }
-
-    private void OnCollisionEnter(Collision collision)
-    {
-        if (collision.gameObject.tag == "Player")
-        {
-            Destroy(gameObject, 0.0f);
-        }
-    }
-}
