--- Animon/Assets/Scripts/CoinSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 70183531daf5f1d4491d0c0d4565c0e5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
