--- Animon/Assets/Texture.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aa1eb2939ade54349b663694a528d2d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
