--- Animon/Assets/Texture/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28747504b90965d4697a6290b76e63cc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
