--- Animon/Assets/Texture/Materials/Albedo_GrassFlower.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 312459a6ebe3d784c9214f6fb289bd9a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
