--- Animon/Assets/Texture/Materials/Grass_Base_Color.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 280abde558d97f542bccfdf6fe989040
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
